backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI & ML
anthropic==0.7.8
sentence-transformers==2.2.2
qdrant-client==1.7.0
numpy==1.24.3
torch==2.1.1
transformers==4.36.2

# Audio Processing
elevenlabs==0.2.26
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
markdown==3.5.1

# Background Jobs
celery==5.3.4
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1